Add Updated Files

+++ app/Http/Controllers/PlayerController.php
@@ -8,6 +8,7 @@ use Illuminate\Support\Str;
 use App\Repository\PlayerRepositoryInterface; 
 use App\Repository\TeamRepositoryInterface;
 use App\Http\Traits\UploadTrait;
+use Illuminate\Support\Facades\Validator;
 
 class PlayerController extends Controller
 { 
@@ -37,18 +38,7 @@ class PlayerController extends Controller
 	public function index()
 	{		
 		$players = $this->playerRepository->getAllPlayers();
-		return view('players.index',compact('players'))->with('i', (request()->input('page', 1) - 1) * 5);
-	}
-	
-	/**
-	* Show the form for creating a new resource.
-	*
-	* @return \Illuminate\Http\Response
-	*/
-	public function create()
-	{	
-		$data = $this->teamRepository->getAllTeams();
-		return view('players.create',compact('data'));
+		return $players;		
 	}
 	
 	/**
@@ -60,28 +50,31 @@ class PlayerController extends Controller
 	public function store(Request $request)
 	{
 		// Form validation
-        request()->validate([
+		$validator = Validator::make($request->all(),[       
 			'firstName' => 'required',
 			'lastName' => 'required',
 			'team_id' => 'required',
 			'playerImageURI' => 'required|mimes:jpg,png|max:2048',
-		]);        
+        ]); 
+		if ($validator->fails()) {          
+            return response()->json(['error'=>$validator->errors()], 422);                        
+        }      
 		$fileName = null;
         // Check if a profile image has been uploaded
-        if ($request->has('playerImageURI')) {
-			$fileName = $this->storeImage($request);
-		}
-		//insert in to database table
-		if( $fileName !='' ) {			
-			$player = ['firstName'=>$request->firstName,'lastName'=>$request->lastName,'team_id'=>$request->team_id,'playerImageURI'=>$fileName];
-			$create = $this->playerRepository->createPlayer($player);
-			if($create !='' ){
-				return redirect()->route('players.index')->with('success','Player created successfully.');
-			}else{		
-				return back()->withErrors(['Player creation failed, Please try again']);
+		if ($request->has('playerImageURI')) {
+			try{
+				$fileName = $this->storeImage($request);
+			}catch(\Exception $e){
+				return response()->json(['error'=>'Internal Server Error'], 500);    
 			}
-		}else{		
-			return back()->withErrors(['Image Upload Failed']);
+		}
+		$player = ['firstName'=>$request->firstName,'lastName'=>$request->lastName,'team_id'=>$request->team_id,'playerImageURI'=>$fileName];
+		//insert in to database table
+		try{
+			$this->playerRepository->createPlayer($player);
+			return response()->json(['success'=>'Player Created Successfully'], 201);   
+		}catch(\Exception $e){
+			return response()->json(['error'=>'Internal Server Error'], 500);    
 		}
 	}
 	
@@ -93,20 +86,13 @@ class PlayerController extends Controller
 	*/
 	public function show(Player $player)
 	{		
-		return view('players.show',compact('player'));
-	}
-	
-	/**
-	* Show the form for editing the specified resource.
-	*
-	* @param  \App\Player  $player
-	* @return \Illuminate\Http\Response
-	*/
-	public function edit(Player $player)
-	{
-		$player['data'] = $this->playerRepository->getAllPlayers();
-		return view('players.edit',compact('player'));
-	}
+		try{
+			$player = $this->playerRepository->getSinglePlayer($player);
+		}catch(ModelNotFoundException $e) {			
+			return response()->json(['error'=>'Resource Not Found'], 422);   
+		}
+		return $player;
+	}	
 	
 	/**
 	* Update the specified resource in storage.
@@ -118,23 +104,39 @@ class PlayerController extends Controller
 	public function update(Request $request, Player $player)
 	{		
 		$save = [];
-		if ($request->hasFile('playerImageURI')) {			
-			request()->validate(['firstName' => 'required','lastName' => 'required','team_id' => 'required','playerImageURI' => 'required|mimes:jpg,png|max:2048',]);
-			//delete previous image			
+		if ($request->hasFile('playerImageURI')) {
+			// Form validation
+			$validator = Validator::make($request->all(),[       
+				'firstName' => 'required',
+				'lastName' => 'required',
+				'team_id' => 'required',
+				'playerImageURI' => 'required|mimes:jpg,png|max:2048',
+			]); 
+			if ($validator->fails()) {          
+				return response()->json(['error'=>$validator->errors()], 422);                        
+			}      
 			$this->deleteOne(public_path('uploads/'), $player->playerImageURI);
 			//upload new image
-			$fileName = $this->storeImage($request);
+			try{
+				$fileName = $this->storeImage($request);
+			}catch(\Exception $e){
+				return response()->json(['error'=>'Internal Server Error'], 500);    
+			}
 			$save['playerImageURI'] = $fileName;
 		}else{
-			request()->validate(['firstName' => 'required','lastName' => 'required','team_id' => 'required']);
+			$validator = Validator::make($request->all(),['firstName' => 'required','lastName' => 'required','team_id' => 'required',]); 
+			if ($validator->fails()) {          
+				return response()->json(['error'=>$validator->errors()], 422);                        
+			} 
 		}
 		$save['firstName'] = $request->firstName;
 		$save['lastName'] = $request->lastName;
 		$save['team_id'] = $request->team_id;
-		if($this->playerRepository->updatePlayer($save, $player->id)) {
-			return redirect()->route('players.index')->with('success','Player updated successfully.');
-		}else{		
-			return back()->withErrors(['Updates failed, Please try again']);
+		try{
+			$this->playerRepository->updatePlayer($save, $player->id);
+			return response()->json(['success'=>'Player Updated Successfully'], 201); 
+		}catch(\Exception $e){
+			return response()->json(['error'=>'Internal Server Error'], 500);    
 		}
 	}
 	
@@ -145,9 +147,13 @@ class PlayerController extends Controller
 	* @return \Illuminate\Http\Response
 	*/
 	public function destroy(Player $player)
-	{
-		$player->delete();
-		return redirect()->route('players.index')->with('success','Player deleted successfully');
+	{		
+		try{
+			$this->playerRepository->deletePlayer($player->id);
+			return response()->json(['success'=>'Player UDeleted Successfully'], 201); 
+		}catch(\Exception $e){
+			return response()->json(['error'=>'Internal Server Error'], 500);    
+		}
 	}
 	
 	/**
