+++ app/Http/Controllers/TeamController.php
@@ -7,6 +7,9 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 use App\Repository\TeamRepositoryInterface;
 use App\Http\Traits\UploadTrait;
+use Illuminate\Support\Facades\Validator;
+//use Request;
+use Illuminate\Support\Facades\Input;
 
 class TeamController extends Controller
 { 
@@ -14,7 +17,7 @@ class TeamController extends Controller
 	private $teamRepository;
 	
 	/**
-	* Display a listing of the resource.
+	* Resource Constructor.
 	*
 	* @return \Illuminate\Http\Response
 	*/
@@ -31,21 +34,15 @@ class TeamController extends Controller
 	*
 	* @return \Illuminate\Http\Response
 	*/
-	public function index()
-	{
-		
-		$teams = $this->teamRepository->getAllTeams();
-		return view('teams.index',compact('teams'))->with('i', (request()->input('page', 1) - 1) * 5);
-	}
-	
-	/**
-	* Show the form for creating a new resource.
-	*
-	* @return \Illuminate\Http\Response
-	*/
-	public function create()
-	{
-		return view('teams.create');
+	public function index(Request $request)
+	{		
+		if($request->has('section'))
+		{			
+			$teams = $this->teamRepository->getAllTeams();
+		}else{
+			$teams = $this->teamRepository->getAllTeamsPaging();			
+		}
+		return $teams;		
 	}
 	
 	/**
@@ -57,26 +54,29 @@ class TeamController extends Controller
 	public function store(Request $request)
 	{
 		// Form validation
-        request()->validate([
+        $validator = Validator::make($request->all(),[       
 			'name' => 'required',
-			'logoURI' => 'required|mimes:jpg,png|max:2048',
-		]);        
+            'logoURI' => 'required|mimes:jpg,png|max:2048',
+        ]);   
+ 
+		if ($validator->fails()) {          
+            return response()->json(['error'=>$validator->errors()], 422);                        
+        } 
 		$fileName = null;
         // Check if a profile image has been uploaded
         if ($request->has('logoURI')) {
-			$fileName = $this->storeImage($request);
-		}
-		//insert in to database table
-		if( $fileName !='' ) {			
-			$team = ['name'=>$request->name,'logoURI'=>$fileName];
-			$create = $this->teamRepository->createTeam($team);
-			if($create !='' ){
-				return redirect()->route('teams.index')->with('success','Team created successfully.');
-			}else{		
-				return back()->withErrors(['Team creation failed, Please try again']);
+			try{
+				$fileName = $this->storeImage($request);
+			}catch(\Exception $e){
+				return response()->json(['error'=>'Internal Server Error'], 500);    
 			}
-		}else{		
-			return back()->withErrors(['Image Upload Failed']);
+		}
+		$team = ['name'=>$request->name,'logoURI'=>$fileName];
+		try{
+			$this->teamRepository->createTeam($team);
+			return response()->json(['success'=>'Team Created Successfully'], 201);   
+		}catch(\Exception $e){
+			return response()->json(['error'=>'Internal Server Error'], 500);    
 		}		
 	}
 	
@@ -88,19 +88,14 @@ class TeamController extends Controller
 	*/
 	public function show(Team $team)
 	{
-		return view('teams.show',compact('team'));
-	}
-	
-	/**
-	* Show the form for editing the specified resource.
-	*
-	* @param  \App\Team  $team
-	* @return \Illuminate\Http\Response
-	*/
-	public function edit(Team $team)
-	{
-		return view('teams.edit',compact('team'));
-	}
+		try{
+			$team = $this->teamRepository->getSingleTeam($team);
+		}catch(ModelNotFoundException $e) {			
+			return response()->json(['error'=>'Resource Not Found'], 422);   
+		}
+		return $team;
+		
+	}	
 	
 	/**
 	* Update the specified resource in storage.
@@ -110,23 +105,37 @@ class TeamController extends Controller
 	* @return \Illuminate\Http\Response
 	*/
 	public function update(Request $request, Team $team)
-	{		
+	{	
 		$save = [];
 		if ($request->hasFile('logoURI')) {			
-			request()->validate(['name' => 'required','logoURI' => 'required|mimes:jpg,png|max:2048',]);
+			$validator = Validator::make($request->all(),[       
+				'name' => 'required',
+				'logoURI' => 'required|mimes:jpg,png|max:2048',
+			]);
+			if ($validator->fails()) {          
+				return response()->json(['error'=>$validator->errors()], 422);                        
+			} 
 			//delete previous image			
 			$this->deleteOne(public_path('uploads/'), $team->logoURI);
 			//upload new image
-			$fileName = $this->storeImage($request);
+			try{
+				$fileName = $this->storeImage($request);
+			}catch(\Exception $e){
+				return response()->json(['error'=>'Internal Server Error'], 500);    
+			}
 			$save['logoURI'] = $fileName;
 		}else{
-			request()->validate(['name' => 'required']);
+			$validator = Validator::make($request->all(),[ 'name' => 'required',]);
+			if ($validator->fails()) {          
+				return response()->json(['error'=>$validator->errors()], 422);                        
+			} 
 		}
 		$save['name'] = $request->name;
-		if($this->teamRepository->updateTeam($save, $team->id)) {
-			return redirect()->route('teams.index')->with('success','Team updated successfully.');
-		}else{		
-			return back()->withErrors(['Updates failed, Please try again']);
+		try{
+			$this->teamRepository->updateTeam($save, $team->id);
+			return response()->json(['success'=>'Team Updated Successfully'], 201); 
+		}catch(\Exception $e){
+			return response()->json(['error'=>'Internal Server Error'], 500);    
 		}
 	}
 	
@@ -138,8 +147,12 @@ class TeamController extends Controller
 	*/
 	public function destroy(Team $team)
 	{
-		$team->delete();
-		return redirect()->route('teams.index')->with('success','Team deleted successfully');
+		try{
+			$this->teamRepository->deleteTeam($team->id);
+			return response()->json(['success'=>'Team UDeleted Successfully'], 201); 
+		}catch(\Exception $e){
+			return response()->json(['error'=>'Internal Server Error'], 500);    
+		}	
 	}
 	
 	/**
@@ -158,5 +171,6 @@ class TeamController extends Controller
         $result = $this->uploadOne($image, $folder, 'public', $name);
 		$fileURL = url('/uploads').'/'.$result;
 		return $fileURL;
+		
 	}
 }
